+++ db/compaction/compaction.cc
@@ -546,16 +546,16 @@ bool Compaction::ShouldFormSubcompactions() const {
 }
 
 uint64_t Compaction::MinInputFileCreationTime() const {
-  uint64_t max_creation_time = 0;
+  uint64_t min_creation_time = port::kMaxUint64;
   for (const auto& file : inputs_[0].files) {
     if (file->fd.table_reader != nullptr &&
         file->fd.table_reader->GetTableProperties() != nullptr) {
       uint64_t creation_time =
           file->fd.table_reader->GetTableProperties()->creation_time;
-      max_creation_time = std::min(max_creation_time, creation_time);
+      min_creation_time = std::min(min_creation_time, creation_time);
     }
   }
-  return max_creation_time;
+  return min_creation_time;
 }
 
 int Compaction::GetInputBaseLevel() const {